requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=8.3.1
opencv-python>=4.5.3
scikit-learn>=0.24.2
matplotlib>=3.4.3
tqdm>=4.62.2
tensorboard>=2.7.0
albumentations>=1.1.0
face-alignment>=1.3.5
dlib>=19.24.0 